requirements.txt
-i https://pypi.org/simple
anyio==3.7.1; python_version >= '3.7'
asgiref==3.8.1
bidict==0.23.1; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
flask[async]==3.1.0; python_version >= '3.9'
flask-socketio==5.5.1; python_version >= '3.6'
h11==0.12.0; python_version >= '3.6'
httpcore==0.15.0; python_version >= '3.7'
httpx==0.23.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
imageio[ffmpeg]==2.37.0; python_version >= '3.9'
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
numpy==2.2.2; python_version >= '3.10'
pillow==11.1.0; python_version >= '3.9'
psutil==6.1.1
python-dotenv==1.0.1; python_version >= '3.8'
python-engineio==4.11.2; python_version >= '3.6'
python-socketio==5.12.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rfc3986[idna2008]==1.5.0
simple-websocket==1.1.0; python_version >= '3.6'
sniffio==1.3.1; python_version >= '3.7'
urllib3==2.3.0; python_version >= '3.9'
utils==1.0.2; python_version >= '3.6'
werkzeug==3.1.3; python_version >= '3.9'
wsproto==1.2.0; python_full_version >= '3.7.0'
youtube-search-python==1.6.6; python_version >= '3.6'
yt-dlp==2025.1.26; python_version >= '3.9'
